backend/requirements.txt
fastapi
uvicorn[standard]==0.23.2
SQLAlchemy
psycopg2-binary==2.9.6
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
pydantic==1.10.12
aiofiles==23.1.0
python-dotenv==1.0.0
pydantic[email]==1.10.12
